pyproject.toml
[project]
name = "mlops-lab3"
version = "0.1.0"
description = "MLOps Lab3 - Transfer Learning with MLFlow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.9.0",
    "click>=8.3.0",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pillow>=12.0.0",
    "pylint>=4.0.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "mlflow>=2.9.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "onnxscript>=0.5.6",
]
